Rename core dir + formatting
