+++ albius.go
@@ -3,14 +3,14 @@ package main
 import (
 	"fmt"
 
-	"github.com/vanilla-os/albius/albius"
+	"github.com/vanilla-os/albius/core"
 )
 
 func main() {
-    disk, err := albius.LocateDisk("/dev/nvme0n1")
-    if err != nil {
-        panic(err)
-    }
+	disk, err := albius.LocateDisk("/dev/nvme0n1")
+	if err != nil {
+		panic(err)
+	}
 
-    fmt.Println(disk)
+	fmt.Println(disk)
 }
