+++ core/util.go
